The level loading has been redesign.
A DynamicMap should now be created which loads a "story" world, which is a list of level files. This allows the Player to remain a single instance and keeps level state when moveing between them.

The level are them created using the LevelFactory. Basic themeing support is also available, it changes the base content path for tiles to looking the the Themes folder.

All other functionality as of Friday has been made to work with the new system.